Adds File System and GUI Interface.
Fairly Standard File System: all coding files go in the source. We could
go even further, and add more depth to the source. But this is a good
start.

Standardizing our GUI classes: Adds Interface for GUIs From now one all
GUIs will have to implement this interface (see MenuPanel). The
interface is basically the minimal required methods for each gui. Anyone
can add a method to the interface however, they will have to add said
method to each GUI class.
