--- MAIN.java
@@ -1,27 +1,0 @@
-/* 
-Name: main 
-Author: Joshua Becker
-Create On: 9/9/15
-Updated On: 9/10/15
-Contributors:
- */
-
-import javax.swing.JFrame; 
-
-public class MAIN
- {
-	/**main
-	* create frame, create panel, display;
-	* J.B.
-	**/
-	public static void main(String[] args)
-    {       
-	JFrame menuFrame = new JFrame("Main Menu");
-	menuFrame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
-	
-    MenuPanel panel = new MenuPanel(menuFrame); 
-    menuFrame.pack();
-	menuFrame.setVisible(true);
-	
-    }
-}
